pyproject.toml
[project]
name = "pandas-agent-core"
version = "0.1.0"
description = "Standalone pandas agent for data analysis using LangChain"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-experimental>=0.0.50",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.20",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "langchain-aws>=0.1.0",
    "boto3>=1.34.0",
    "flask>=3.0.0",
    "openpyxl>=3.1.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
